move walkerlog to driver directory
