+++ src/Estimators/CMakeLists.txt
@@ -26,8 +26,6 @@ set(QMCEST_SRC
     EstimatorManagerBase.cpp
     EstimatorManagerNew.cpp
     EstimatorManagerCrowd.cpp
-    WalkerLogCollector.cpp
-    WalkerLogManager.cpp
     CollectablesEstimator.cpp
     OperatorEstBase.cpp
     SpinDensityNew.cpp
