+++ src/Estimators/tests/CMakeLists.txt
@@ -29,7 +29,6 @@ set(SRCS
     test_manager.cpp
     test_EstimatorManagerNew.cpp
     test_trace_manager.cpp
-    test_WalkerLogCollector.cpp
     EstimatorTesting.cpp
     GenerateRandomParticleSets.cpp
     test_SpinDensityInput.cpp
