+++ src/QMCDrivers/WalkerLogBuffer.h
