+++ src/QMCDrivers/WalkerLogCollector.cpp
