+++ src/QMCDrivers/WalkerLogCollector.h
