+++ src/QMCDrivers/WalkerLogManager.cpp
