+++ src/QMCDrivers/WalkerLogManager.h
