+++ src/QMCDrivers/tests/CMakeLists.txt
@@ -20,6 +20,7 @@ set(UTEST_HDF_INPUT ${qmcpack_SOURCE_DIR}/tests/solids/diamondC_1x1x1_pp/pwscf.p
 maybe_symlink(${UTEST_HDF_INPUT} ${UTEST_DIR}/diamondC_1x1x1.pwscf.h5)
 
 set(DRIVER_TEST_SRC
+    test_WalkerLogCollector.cpp
     test_TauParams.cpp
     test_vmc.cpp
     test_dmc.cpp
